modifying some user/group deletion stuff
git-svn-id: http://svn.concrete5.org/svn/concrete5@292 b0551a0c-1e16-4222-a7d5-975db1aca215

Former-commit-id: 3c8c639918838e46a2d4147f92bd169bb0df7e93

+++ trunk/web/concrete/models/groups.php
@@ -246,6 +246,17 @@
 		}
 
 		function delete(){
+			// we will NOT let you delete the required groups
+			if ($this->gID == REGISTERED_GROUP_ID || $this->gID == GUEST_GROUP_ID) {
+				return false;
+			}
+			
+			// run any internal event we have for group deletion
+			$ret = Events::fire('on_group_delete', $this);
+			if ($ret < 0) {
+				return false;
+			}
+			
 			$db = Loader::db(); 
 			$r = $db->query("DELETE FROM UserGroups WHERE gID = ?",array(intval($this->gID)) );
 			$r = $db->query("DELETE FROM Groups WHERE gID = ?",array(intval($this->gID)) );
